Added event off chp.
Javascript:
Added function to check for 0 or negative hp. If true page closes as creature is dead.
Added safety check to prevent page closure if max hp isn't setup.
Added function to take + or - statements and add them to the current hp.

HTML:
Changed all inputs from numbers to tel to remove annoying +|- and to make +3 a legal statement.